Move tests to be close to the code
