+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "carpark-simulator",
+  "name": "email-editor",
   "version": "1.0.0",
-  "description": "This application is a simulation of a robot operated bus moving in a carpark",
+  "description": "This application is an interaction email builder",
   "main": "src/index.js",
   "repository": "https://github.com/sunnyniu/email-editor",
   "author": "Sunny Niu",
@@ -17,7 +17,7 @@
   },
   "jest": {
     "setupFiles": [
-      "./tests/setup.js"
+      "./setupTests.js"
     ],
     "reporters": [
       "default",
@@ -55,11 +55,14 @@
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-react-hooks": "^1.7.0",
+    "html-webpack-plugin": "^3.2.0",
     "jest": "^24.9.0",
     "path": "^0.12.7",
     "prettier": "^1.19.1",
     "redux-devtools-extension": "^2.13.8",
     "redux-mock-store": "^1.5.4",
+    "redux-saga-test-plan": "^4.0.0-rc.3",
+    "supertest": "^4.0.2",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.10.1"
@@ -68,21 +71,15 @@
     "@babel/runtime": "^7.7.7",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "html-webpack-plugin": "^3.2.0",
     "isomorphic-fetch": "^2.2.1",
-    "jest-fetch-mock": "^3.0.3",
     "lodash": "^4.17.15",
-    "nock": "^12.0.3",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.3",
-    "react-test-renderer": "^16.13.1",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3",
-    "redux-thunk": "^2.3.0",
     "styled-components": "^4.4.1",
-    "styled-css-grid": "^1.2.1",
-    "supertest": "^4.0.2"
+    "styled-css-grid": "^1.2.1"
   }
 }
