--- tests/apis.test.js
@@ -1,5 +1,0 @@
-describe('apis tests', () => {
-  it('get xml path return a response object with correct paths', () => {
-    expect(1).toEqual(1);
-  });
-});
