--- tests/setup.js
@@ -1,8 +1,0 @@
-import enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-require('jest-fetch-mock').enableMocks();
-
-enzyme.configure({
-  adapter: new Adapter(),
-});
